requirements.txt
Django==1.11.29
macholib==1.5.1
python-dateutil==1.5
pytz==2013.7
six==1.12.0
psycopg2-binary>=2.8
djangorestframework==3.9.4
